# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateEventRequest:
    properties:
      capacity:
        minimum: 1
        type: integer
      description:
        maxLength: 1000
        type: string
      end_time:
        type: string
      location:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      start_time:
        type: string
    required:
    - capacity
    - end_time
    - location
    - name
    - start_time
    type: object
...
  dto.CreateMerchantRequest:
    properties:
      description:
        maxLength: 500
        type: string
      email:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        type: string
      phone:
        type: string
    required:
    - email
    - name
    - password
    - phone
    type: object
...
  dto.EventResponse:
    properties:
      active:
        type: boolean
      capacity:
        type: integer
      created_at:
        type: string
      description:
        type: string
      end_time:
        type: string
      event_id:
        type: string
      location:
        type: string
      merchant_id:
        type: string
      name:
        type: string
      start_time:
        type: string
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      expires_at:
        description: unix timestamp
        type: integer
      token:
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
...
  dto.MerchantLoginResponse:
    properties:
      expires_at:
        type: integer
      merchant:
        allOf:
        - $ref: '#/definitions/dto.MerchantResponse'
        description: ← named merchant, not user
      token:
        type: string
    type: object
  dto.MerchantRepLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.MerchantRepLoginResponse:
    properties:
      expires_at:
        type: integer
      rep:
        $ref: '#/definitions/dto.MerchantRepResponse'
      token:
        type: string
    type: object
  dto.MerchantRepResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      merchant_id:
        type: string
      merchant_rep_id:
        type: string
      name:
        type: string
      phone:
        type: string
      role:
        type: string
    type: object
  dto.MerchantResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      email:
        type: string
      merchant_id:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
  dto.PaginatedEventsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.EventResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
  dto.UpdateEventRequest:
    properties:
      active:
        description: pointer — false is a valid value
        type: boolean
      capacity:
        minimum: 1
        type: integer
      description:
        maxLength: 1000
        type: string
      end_time:
        type: string
      location:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      start_time:
        type: string
    type: object
...
  dto.UpdateMerchantRequest:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      phone:
        type: string
    type: object
...
  dto.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
host: ticketfair.localhost
info:
  contact:
    email: support@ticketfair.com
    name: TicketFair Support
  description: TicketFair event ticketing platform API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: TicketFair API
  version: "1.0"
paths:
  /merchant/events/{id}:
    put:
      consumes:
      - application/json
      description: Update an event under the authenticated merchant
      parameters:
      - description: Event UUID
        in: path
        name: id
        required: true
        type: string
      - description: Updated event data
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EventResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      summary: Update an event
      tags:
      - Events
  /merchant/events/new:
    post:
      consumes:
      - application/json
      description: Create a new event under the authenticated merchant
      parameters:
      - description: Event data
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/dto.CreateEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.EventResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      summary: Create a new event
      tags:
      - Events
...
  /merchant/update:
    put:
      consumes:
      - application/json
      description: Update the authenticated merchant's details
      parameters:
      - description: Updated merchant data
        in: body
        name: merchant
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateMerchantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MerchantResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      summary: Update a merchant
      tags:
      - Merchants
...
  /public/auth/client/login:
    post:
      consumes:
      - application/json
      description: Authenticate a client and return a JWT token
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Client login
      tags:
      - Auth
...
  /public/auth/merchant/login:
    post:
      consumes:
      - application/json
      description: Authenticate a merchant and return a JWT token
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MerchantLoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Merchant login
      tags:
      - Auth
...
  /public/auth/rep/login:
    post:
      consumes:
      - application/json
      description: Authenticate a merchant rep and return a JWT with role and merchant
        context
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dto.MerchantRepLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MerchantRepLoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Merchant representative login
      tags:
      - Auth
  /public/events:
    get:
      description: Retrieve a paginated list of active upcoming events
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedEventsResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all active events
      tags:
      - Events
  /public/events/{id}:
    get:
      description: Retrieve a single event by its UUID
      parameters:
      - description: Event UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EventResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get an event by ID
      tags:
      - Events
...
  /public/merchant/register:
    post:
      consumes:
      - application/json
      description: Create a new merchant account
      parameters:
      - description: Merchant data
        in: body
        name: merchant
        required: true
        schema:
          $ref: '#/definitions/dto.CreateMerchantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.MerchantResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      summary: Create a new merchant
      tags:
      - Merchants
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"